Fix Android build on non-Linux platforms
The `cfg(target_os)` for the build script contains the name of the OS on which the build takes place (the OS of the host machine). Because of this, the build of the Android port on Windows is broken, as an attempt to add the Windows resource file fails. Probably the build for Windows on Linux, for Linux on Windows, etc., is also broken.
Therefore, in the build script, we must rely only on the values of the `TARGET` environment variable, which contains the target triple of the platform for which the build is performing. `cfg(target_os)` should only be used to include platform-specific dependencies (e.g. `winres` for Windows).